+++ viewport.js
@@ -17,12 +17,28 @@ async function observe(el, intersecting = true) {
 		} else {
 			if (! (observer instanceof IntersectionObserver)) {
 				observer = new IntersectionObserver((entries, observer) => {
-					entries.forEach(({ target, isIntersecting }) => {
+					entries.forEach(({
+						target,
+						isIntersecting,
+						boundingClientRect,
+						intersectionRatio,
+						intersectionRect = null,
+						rootBounds,
+						time,
+					}) => {
 						if (isIntersecting) {
 							const prom = intEls.get(target);
 
 							if (prom instanceof Function) {
-								prom(target);
+								prom({
+									target,
+									isIntersecting,
+									boundingClientRect,
+									intersectionRatio,
+									intersectionRect,
+									rootBounds,
+									time,
+								});
 								intEls.delete(target);
 
 								if (! unIntEls.has(target)) {
@@ -33,7 +49,15 @@ async function observe(el, intersecting = true) {
 							const prom = unIntEls.get(target);
 
 							if (prom instanceof Function) {
-								prom(target);
+								prom({
+									target,
+									isIntersecting,
+									boundingClientRect,
+									intersectionRatio,
+									intersectionRect,
+									rootBounds,
+									time,
+								});
 								unIntEls.delete(target);
 
 								if (! intEls.has(target)) {
@@ -70,25 +94,49 @@ export function getViewportBounds() {
 	];
 }
 
-export async function whenInViewport(el) {
-	if (typeof el === 'string') {
-		return whenInViewport(document.querySelector(el));
-	} else if (! (el instanceof Element)) {
+export async function whenInViewport(target) {
+	if (typeof target === 'string') {
+		return whenInViewport(document.querySelector(target));
+	} else if (! (target instanceof Element)) {
 		throw new Error('Invalid element or selector');
-	} else if (isInViewport(el)) {
-		return el;
+	} else if (isInViewport(target)) {
+		return {
+			target,
+			isIntersecting: true,
+			boundingClientRect: target.getBoundingClientRect(),
+			intersectionRatio: NaN,
+			/**
+			 * DOMRect is poorly supported and calculating value may be impractical,
+			 * so just return an object with all keys set and values of NaN
+			 */
+			intersectionRect: { x: NaN, y: NaN, height: NaN, width: NaN, top: NaN, bttom: NaN, left: NaN, right: NaN},
+			rootBounds: null,
+			time: 0,
+		};
 	} else {
-		return await observe(el, true);
+		return await observe(target, true);
 	}
 }
 
-export async function whenNotInViewport(el) {
-	if (typeof el === 'string') {
-		return whenNotInViewport(document.querySelector(el));
-	} else if (! isInViewport(el)) {
-		return el;
+export async function whenNotInViewport(target) {
+	if (typeof target === 'string') {
+		return await whenNotInViewport(document.querySelector(target));
+	} else if (! isInViewport(target)) {
+		return {
+			target,
+			isIntersecting: false,
+			boundingClientRect: target.getBoundingClientRect(),
+			intersectionRatio: 0,
+			/**
+			 * DOMRect is poorly supported, but we do know that zero area is intersecting,
+			 * so set all keys with values of 0
+			 */
+			intersectionRect: { x: 0, y: 0, height: 0, width: 0, top: 0, bttom: 0, left: 0, right: 0},
+			rootBounds: null,
+			time: 0,
+		};
 	} else {
-		return await observe(el, false);
+		return await observe(target, false);
 	}
 }
 
@@ -100,16 +148,18 @@ export async function whenInViewportFor(el, ms = 1000) {
 	} else {
 		return await new Promise(async resolve => {
 			let done = false;
+			let retVal = null;
 
 			while (! done) {
-				await whenInViewport(el);
+				retVal = await whenInViewport(el);
 				const timer = setTimeout(() => {
 					done = true;
 					resolve(el);
-				}, ms);
+				}, Math.max(ms, 1)) || 1;
 
 				await whenNotInViewport(el).then(() => clearTimeout(timer));
 			}
+			return retVal;
 		});
 	}
 }